Rename EditBoardHeader til BoardLinkActions
